Fix: #256 Long으로 변환하던 elderId 타입 변환 제거

+++ app/src/main/java/com/konkuk/medicarecall/ui/feature/home/viewmodel/HomeViewModel.kt
@@ -116,7 +116,7 @@ class HomeViewModel(
             }
             val elderIdMap = elderIdRepository.getElderIds()
             _elderInfoList.value = elderIdMap.map {
-                ElderInfo(elderId = it.key.toLong(), name = it.value)
+                ElderInfo(elderId = it.key, name = it.value)
             }
             val restoredId = savedStateHandle.get<Long?>(KEY_SELECTED_ELDER_ID)
             if (restoredId != null && _elderInfoList.value.any { it.elderId == restoredId }) {
